refactor(headless): dedupe run-script helpers and extract TaskRunner interface
Promote the copy-adapted Harbor A/B run-script helpers into shared TS
modules and give the task-runner seam an explicit, provider-neutral
interface, without changing any runtime behavior.

- envPath and envIds move to headless-run-env.ts; selectTasksByIds moves
  to fixed-prompt-task-source.ts (parameterized to cover the labeled,
  duplicate-tolerant runtime-policy variant).
- The resume-safe fingerprint trio (subject/toolchain/task-source) moves
  from run-prompt-ab.mjs into experiment-fingerprint.ts, ported
  byte-identically; run-prompt-ab, run-harness-ab, and run-runtime-policy-ab
  now import it instead of reverse-importing the .mjs. The parallel
  prompt-optimization fingerprints are deliberately left untouched: they
  are semantically distinct and merging them would change
  prompt-optimization resume identity.
- DEEPSEEK_V4_FLASH_PRICING moves to deepseek-pricing.ts, keeping the
  stricter fail-loud field validation, imported via a package-local
  subpath so it stays out of the generic public API.
- Extract a provider-neutral TaskRunner/TaskRunInput/TaskRunOutput
  interface from HarborTaskRunner; createHarborTaskRunner remains the sole
  implementation and now returns TaskRunner. Harbor-named aliases are kept
  for existing call sites and the new interface is exported from index.ts.

Refs #1342
